requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.10.0
python-multipart==0.0.6
httpx>=0.25.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
duckduckgo-search==6.3.0
openai>=1.50.0